requirements.txt
torch>=2.2.0
torchvision>=0.17.0
scikit-learn>=1.4.0
pillow>=10.0.0
kagglehub>=0.3.0
huggingface_hub>=0.27.0
numpy>=1.26.0
